Initial commit: Filtered the umrt-ros to get the drivetrain stuff,  example_2 and ros2_control_demo_description, and made it it's own package similar to umrt-arm-ros-firmware, then organized it so that the project name would be changed from ros2_control_demo_descirption to umrt-drivetrain-ros. TBD if this works havent tested it.
