+++ CMakeLists.txt
@@ -1,20 +1,24 @@
 cmake_minimum_required(VERSION 3.16)
-project(********** Fill in project name **********)
+project(umrt-drivetrain-ros)
 
 set(CMAKE_CXX_STANDARD 14)
 
-if(CMAKE_CXX_COMPILER_ID MATCHES "(GNU|Clang)")
-  add_compile_options(-Wall -Wextra -fPIC)
-endif()
+# set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
+# set(CMAKE_CXX_STANDARD_REQUIRED On)
 
-# Fixes intellisense not finding ros2 headers https://youtrack.jetbrains.com/issue/CPP-29747/Certain-ROS2-package-headers-missing-from-Intellisense-when-using-a-Docker-toolchain
-include_directories(SYSTEM /opt/ros/$ENV{ROS_DISTRO}/include)
+if(CMAKE_CXX_COMPILER_ID MATCHES "(GNU|Clang)")
+  add_compile_options(-Wall -Wextra)
+endif()
 
 # find dependencies
 set(THIS_PACKAGE_INCLUDE_DEPENDS
-        rclcpp
-        rclcpp_lifecycle
-        )
+  hardware_interface
+  pluginlib
+  rclcpp
+  rclcpp_lifecycle
+  ros2_socketcan
+  ament_cmake
+)
 
 # find dependencies
 find_package(ament_cmake REQUIRED)
@@ -22,19 +26,52 @@ foreach(Dependency IN ITEMS ${THIS_PACKAGE_INCLUDE_DEPENDS})
   find_package(${Dependency} REQUIRED)
 endforeach()
 
-add_executable(${PROJECT_NAME}
-    src/project_name.cpp
-    src/example_node.cpp
+# Setup ros2_control_demo_description
+install(
+        DIRECTORY hardware/include/ros2_control_demo_description/diffbot/urdf hardware/include/ros2_control_demo_description/diffbot/rviz
+        DESTINATION share/${PROJECT_NAME}/ros2_control_demo_description/diffbot
 )
-target_compile_features(${PROJECT_NAME} PUBLIC cxx_std_17)
-target_include_directories(${PROJECT_NAME} PUBLIC
-        $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
-        $<INSTALL_INTERFACE:include/${PROJECT_NAME}>
-        )
+# EXAMPLE - we name them per year?
+# install(
+#         DIRECTORY hardware/include/ros2_control_demo_description/prairie_pioneer/urdf hardware/include/ros2_control_demo_description/prairie_pioneer/rviz
+#         DESTINATION share/${PROJECT_NAME}/ros2_control_demo_description/prairie_pioneer
+# )
+
+## COMPILE
+add_library(
+  umrt-drivetrain-ros
+  SHARED
+  hardware/diffbot_system.cpp
+  hardware/I2CPeripheral.cpp
+  hardware/PCA9685.cpp
+)
+
+target_compile_features(umrt-drivetrain-ros PUBLIC cxx_std_17)
+target_include_directories(umrt-drivetrain-ros PUBLIC
+$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/hardware/include>
+$<INSTALL_INTERFACE:include/umrt-drivetrain-ros>
+)
+
+target_link_libraries(umrt-drivetrain-ros
+        PUBLIC
+        $(THIS_PACKAGE_INCLUDE_DEPENDS)
+)
+
 ament_target_dependencies(
-        ${PROJECT_NAME} PUBLIC
-        ${THIS_PACKAGE_INCLUDE_DEPENDS}
+  umrt-driverain-ros PUBLIC
+  ${THIS_PACKAGE_INCLUDE_DEPENDS}
 )
+
+# GOT FROM umrt-arm-ros-firmware, might need it in the future if we are seperating the drivetrain libraries for CAN stuff
+# ament_target_dependencies doesn't use namespaces when linking, so we need to link umrt-arm-firmware-lib ourself
+# target_link_libraries(umrt-arm-ros-firmware PRIVATE umrt-arm-firmware-lib::umrt-arm-firmware-lib)
+
+# Causes the visibility macros to use dllexport rather than dllimport,
+# which is appropriate when building the dll but not consuming it.
+target_compile_definitions(${PROJECT_NAME} PRIVATE "umrt-drivetrain-ros_BUILDING_DLL")
+
+# Export hardware plugins
+pluginlib_export_plugin_description_file(hardware_interface umrt-drivetrain-ros.xml)
 
 # Parse version number from the file
 file(READ "version" version_input)
@@ -48,14 +85,32 @@ configure_file(
 )
 
 # INSTALL
-install(TARGETS
-        ${PROJECT_NAME}
-        DESTINATION lib/${PROJECT_NAME}
-        )
-
 install(
-        DIRECTORY launch
-        DESTINATION share/${PROJECT_NAME}/
+  DIRECTORY hardware/include/
+  DESTINATION include/umrt-drivetrain-ros
+)
+install(
+  DIRECTORY description/launch description/ros2_control description/urdf
+  DESTINATION share/umrt-drivetrain-ros
+)
+install(
+  DIRECTORY bringup/launch bringup/config
+  DESTINATION share/umrt-drivetrain-ros
+)
+install(TARGETS umrt-drivetrain-ros
+  EXPORT export_umrt-drivetrain-ros
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin
 )
 
+if(BUILD_TESTING)
+  find_package(ament_cmake_pytest REQUIRED)
+  ament_add_pytest_test(example_2_urdf_xacro test/test_urdf_xacro.py)
+  ament_add_pytest_test(view_example_2_launch test/test_view_robot_launch.py)
+endif()
+
+## EXPORTS
+ament_export_targets(export_umrt-drivetrain-ros HAS_LIBRARY_TARGET)
+ament_export_dependencies(${THIS_PACKAGE_INCLUDE_DEPENDS})
 ament_package()
