+++ description/launch/view_robot.launch.py
@@ -69,7 +69,7 @@ def generate_launch_description():
             PathJoinSubstitution([FindExecutable(name="xacro")]),
             " ",
             PathJoinSubstitution(
-                [FindPackageShare("ros2_control_demo_example_2"), "urdf", description_file]
+                [FindPackageShare("umrt-drivetrain-ros"), "urdf", description_file]
             ),
             " ",
             "prefix:=",
