+++ doc/userdoc.rst
