--- example_2/CMakeLists.txt
@@ -1,88 +1,0 @@
-cmake_minimum_required(VERSION 3.16)
-project(ros2_control_demo_example_2 LANGUAGES CXX)
-set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
-set(CMAKE_CXX_STANDARD_REQUIRED On)
-
-if(CMAKE_CXX_COMPILER_ID MATCHES "(GNU|Clang)")
-  add_compile_options(-Wall -Wextra)
-endif()
-
-# find dependencies
-set(THIS_PACKAGE_INCLUDE_DEPENDS
-  hardware_interface
-  pluginlib
-  rclcpp
-  rclcpp_lifecycle
-  ros2_socketcan
-)
-
-# find dependencies
-find_package(ament_cmake REQUIRED)
-foreach(Dependency IN ITEMS ${THIS_PACKAGE_INCLUDE_DEPENDS})
-  find_package(${Dependency} REQUIRED)
-endforeach()
-
-
-## COMPILE
-add_library(
-  ros2_control_demo_example_2
-  SHARED
-  hardware/diffbot_system.cpp
-  hardware/I2CPeripheral.cpp
-  hardware/PCA9685.cpp
-)
-
-target_compile_features(ros2_control_demo_example_2 PUBLIC cxx_std_17)
-target_include_directories(ros2_control_demo_example_2 PUBLIC
-$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/hardware/include>
-$<INSTALL_INTERFACE:include/ros2_control_demo_example_2>
-)
-
-target_link_libraries(ros2_control_demo_example_2
-        PUBLIC
-        i2c
-)
-
-ament_target_dependencies(
-  ros2_control_demo_example_2 PUBLIC
-  ${THIS_PACKAGE_INCLUDE_DEPENDS}
-)
-
-# Causes the visibility macros to use dllexport rather than dllimport,
-# which is appropriate when building the dll but not consuming it.
-target_compile_definitions(${PROJECT_NAME} PRIVATE "ROS2_CONTROL_DEMO_EXAMPLE_2_BUILDING_DLL")
-
-# Export hardware plugins
-pluginlib_export_plugin_description_file(hardware_interface ros2_control_demo_example_2.xml)
-
-# INSTALL
-install(
-  DIRECTORY hardware/include/
-  DESTINATION include/ros2_control_demo_example_2
-)
-install(
-  DIRECTORY description/launch description/ros2_control description/urdf
-  DESTINATION share/ros2_control_demo_example_2
-)
-install(
-  DIRECTORY bringup/launch bringup/config
-  DESTINATION share/ros2_control_demo_example_2
-)
-install(TARGETS ros2_control_demo_example_2
-  EXPORT export_ros2_control_demo_example_2
-  ARCHIVE DESTINATION lib
-  LIBRARY DESTINATION lib
-  RUNTIME DESTINATION bin
-)
-
-if(BUILD_TESTING)
-  find_package(ament_cmake_pytest REQUIRED)
-
-  ament_add_pytest_test(example_2_urdf_xacro test/test_urdf_xacro.py)
-  ament_add_pytest_test(view_example_2_launch test/test_view_robot_launch.py)
-endif()
-
-## EXPORTS
-ament_export_targets(export_ros2_control_demo_example_2 HAS_LIBRARY_TARGET)
-ament_export_dependencies(${THIS_PACKAGE_INCLUDE_DEPENDS})
-ament_package()
