+++ hardware/include/umrt-drivetrain-ros/diffbot_system.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef ROS2_CONTROL_DEMO_EXAMPLE_2__DIFFBOT_SYSTEM_HPP_
-#define ROS2_CONTROL_DEMO_EXAMPLE_2__DIFFBOT_SYSTEM_HPP_
+#ifndef UMRT_DRIVETRAIN_ROS__DIFFBOT_SYSTEM_HPP_
+#define UMRT_DRIVETRAIN_ROS__DIFFBOT_SYSTEM_HPP_
 
 #include <memory>
 #include <string>
@@ -30,11 +30,11 @@
 #include "rclcpp_lifecycle/node_interfaces/lifecycle_node_interface.hpp"
 #include "rclcpp_lifecycle/state.hpp"
 
-#include "ros2_control_demo_example_2/visibility_control.h"
+#include "umrt-drivetrain-ros/visibility_control.h"
 
 #include <ros2_socketcan/socket_can_sender.hpp>
 
-#include "ros2_control_demo_example_2/wheel_info.hpp"
+#include "umrt-drivetrain-ros/wheel_info.hpp"
 
 // This is a hacky workaround to get velocity commands working on the robot. we
 // are provided per-wheel angular velocitites, which we convert to a +/- 1 m/s
@@ -49,33 +49,33 @@ class DiffBotSystemHardware : public hardware_interface::SystemInterface
 public:
   RCLCPP_SHARED_PTR_DEFINITIONS(DiffBotSystemHardware);
 
-  ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
+  UMRT_DRIVETRAIN_ROS_PUBLIC
   hardware_interface::CallbackReturn on_init(
     const hardware_interface::HardwareInfo & info) override;
 
-  ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
+  UMRT_DRIVETRAIN_ROS_PUBLIC
   std::vector<hardware_interface::StateInterface> export_state_interfaces() override;
 
-  ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
+  UMRT_DRIVETRAIN_ROS_PUBLIC
   std::vector<hardware_interface::CommandInterface> export_command_interfaces() override;
 
-  ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
+  UMRT_DRIVETRAIN_ROS_PUBLIC
   hardware_interface::CallbackReturn on_activate(
     const rclcpp_lifecycle::State & previous_state) override;
 
-  ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
+  UMRT_DRIVETRAIN_ROS_PUBLIC
   hardware_interface::CallbackReturn on_deactivate(
     const rclcpp_lifecycle::State & previous_state) override;
 
-  ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
+  UMRT_DRIVETRAIN_ROS_PUBLIC
   hardware_interface::return_type read(
     const rclcpp::Time & time, const rclcpp::Duration & period) override;
 
-  ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
+  UMRT_DRIVETRAIN_ROS_PUBLIC
   hardware_interface::return_type write(
     const rclcpp::Time & time, const rclcpp::Duration & period) override;
 
-  ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
+  UMRT_DRIVETRAIN_ROS_PUBLIC
   hardware_interface::CallbackReturn on_error(
     const rclcpp_lifecycle::State &previous_state) override;
 
@@ -103,4 +103,4 @@ private:
 
 }  // namespace ros2_control_demo_example_2
 
-#endif  // ROS2_CONTROL_DEMO_EXAMPLE_2__DIFFBOT_SYSTEM_HPP_
+#endif  // UMRT_DRIVETRAIN_ROS__DIFFBOT_SYSTEM_HPP_
