+++ hardware/include/umrt-drivetrain-ros/visibility_control.h
@@ -19,38 +19,38 @@
  * library cannot have, but the consuming code must have inorder to link.
  */
 
-#ifndef ROS2_CONTROL_DEMO_EXAMPLE_2__VISIBILITY_CONTROL_H_
-#define ROS2_CONTROL_DEMO_EXAMPLE_2__VISIBILITY_CONTROL_H_
+#ifndef UMRT_DRIVETRAIN_ROS__VISIBILITY_CONTROL_H_
+#define UMRT_DRIVETRAIN_ROS__VISIBILITY_CONTROL_H_
 
 // This logic was borrowed (then namespaced) from the examples on the gcc wiki:
 //     https://gcc.gnu.org/wiki/Visibility
 
 #if defined _WIN32 || defined __CYGWIN__
 #ifdef __GNUC__
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_EXPORT __attribute__((dllexport))
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_IMPORT __attribute__((dllimport))
+#define UMRT_DRIVETRAIN_ROS_EXPORT __attribute__((dllexport))
+#define UMRT_DRIVETRAIN_ROS_IMPORT __attribute__((dllimport))
 #else
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_EXPORT __declspec(dllexport)
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_IMPORT __declspec(dllimport)
+#define UMRT_DRIVETRAIN_ROS_EXPORT __declspec(dllexport)
+#define UMRT_DRIVETRAIN_ROS_IMPORT __declspec(dllimport)
 #endif
-#ifdef ROS2_CONTROL_DEMO_EXAMPLE_2_BUILDING_DLL
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC ROS2_CONTROL_DEMO_EXAMPLE_2_EXPORT
+#ifdef UMRT_DRIVETRAIN_ROS_BUILDING_DLL
+#define UMRT_DRIVETRAIN_ROS_PUBLIC UMRT_DRIVETRAIN_ROS_EXPORT
 #else
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC ROS2_CONTROL_DEMO_EXAMPLE_2_IMPORT
+#define UMRT_DRIVETRAIN_ROS_PUBLIC UMRT_DRIVETRAIN_ROS_IMPORT
 #endif
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC_TYPE ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_LOCAL
+#define UMRT_DRIVETRAIN_ROS_PUBLIC_TYPE UMRT_DRIVETRAIN_ROS_PUBLIC
+#define UMRT_DRIVETRAIN_ROS_LOCAL
 #else
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_EXPORT __attribute__((visibility("default")))
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_IMPORT
+#define UMRT_DRIVETRAIN_ROS_EXPORT __attribute__((visibility("default")))
+#define UMRT_DRIVETRAIN_ROS_IMPORT
 #if __GNUC__ >= 4
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC __attribute__((visibility("default")))
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_LOCAL __attribute__((visibility("hidden")))
+#define UMRT_DRIVETRAIN_ROS_PUBLIC __attribute__((visibility("default")))
+#define UMRT_DRIVETRAIN_ROS_LOCAL __attribute__((visibility("hidden")))
 #else
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_LOCAL
+#define UMRT_DRIVETRAIN_ROS_PUBLIC
+#define UMRT_DRIVETRAIN_ROS_LOCAL
 #endif
-#define ROS2_CONTROL_DEMO_EXAMPLE_2_PUBLIC_TYPE
+#define UMRT_DRIVETRAIN_ROS_PUBLIC_TYPE
 #endif
 
-#endif  // ROS2_CONTROL_DEMO_EXAMPLE_2__VISIBILITY_CONTROL_H_
+#endif  // UMRT_DRIVETRAIN_ROS__VISIBILITY_CONTROL_H_
