--- launch/example.yaml
@@ -1,3 +1,0 @@
-umrt_example_node:
-  ros__parameters:
-    example_param: 0
