+++ mainpage.dox
@@ -1,5 +1,5 @@
 /** \mainpage
-This library/executable/project implements XYZ functionality for the University of Manitoba Robotics Team's
+This library/executable/project implements drivetrain functionality for the University of Manitoba Robotics Team's
 rover/robotic arm.
 <p>
 This includes:
@@ -10,5 +10,5 @@ This includes:
 A small program to show how to set up a project's file tree.
 
 <hr>
-The Doxygen tagfile for this documentation is available <a href="********** project name **********.tag.xml">here</a>.
+The Doxygen tagfile for this documentation is available <a href="umrt-drivetrain-ros.tag.xml">here</a>.
 */
