+++ test/test_urdf_xacro.py
@@ -29,26 +29,49 @@
 # Author: Lukas Sackewitz
 
 import os
-import pytest
+import shutil
+import subprocess
+import tempfile
 
 from ament_index_python.packages import get_package_share_directory
-from launch import LaunchDescription
-from launch.actions import IncludeLaunchDescription
-from launch.launch_description_sources import PythonLaunchDescriptionSource
-from launch_testing.actions import ReadyToTest
 
 
-# Executes the given launch file and checks if all nodes can be started
-@pytest.mark.rostest
-def generate_test_description():
-    launch_include = IncludeLaunchDescription(
-        PythonLaunchDescriptionSource(
-            os.path.join(
-                get_package_share_directory("ros2_control_demo_example_2"),
-                "launch/view_robot.launch.py",
-            )
-        ),
-        launch_arguments={"gui": "true"}.items(),
+def test_urdf_xacro():
+    # General Arguments
+    description_package = "umrt-drivetrain-ros"
+    description_file = "diffbot.urdf.xacro"
+
+    description_file_path = os.path.join(
+        get_package_share_directory(description_package), "urdf", description_file
     )
 
-    return LaunchDescription([launch_include, ReadyToTest()])
+    (_, tmp_urdf_output_file) = tempfile.mkstemp(suffix=".urdf")
+
+    # Compose `xacro` and `check_urdf` command
+    xacro_command = (
+        f"{shutil.which('xacro')}" f" {description_file_path}" f" > {tmp_urdf_output_file}"
+    )
+    check_urdf_command = f"{shutil.which('check_urdf')} {tmp_urdf_output_file}"
+
+    # Try to call processes but finally remove the temp file
+    try:
+        xacro_process = subprocess.run(
+            xacro_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True
+        )
+
+        assert xacro_process.returncode == 0, " --- XACRO command failed ---"
+
+        check_urdf_process = subprocess.run(
+            check_urdf_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True
+        )
+
+        assert (
+            check_urdf_process.returncode == 0
+        ), "\n --- URDF check failed! --- \nYour xacro does not unfold into a proper urdf robot description. Please check your xacro file."
+
+    finally:
+        os.remove(tmp_urdf_output_file)
+
+
+if __name__ == "__main__":
+    test_urdf_xacro()
