+++ umrt-drivetrain-ros.xml
@@ -1,6 +1,6 @@
-<library path="ros2_control_demo_example_2">
-  <class name="ros2_control_demo_example_2/DiffBotSystemHardware"
-         type="ros2_control_demo_example_2::DiffBotSystemHardware"
+<library path="umrt-drivetrain-ros">
+  <class name="umrt-drivetrain-ros/DiffBotSystemHardware"
+         type="umrt-drivetrain-ros::DiffBotSystemHardware"
          base_class_type="hardware_interface::SystemInterface">
     <description>
       The ros2_control DiffBot example using a system hardware interface-type. It uses velocity command and position state interface. The example is the starting point to implement a hardware interface for differential-drive mobile robots.
